% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emsats.R
\name{EMSATS}
\alias{EMSATS}
\title{PCA based adapitve SNP-set association test of multiple traits using GWAS summary statistics}
\usage{
EMSATS(Z, Sig, R, rho = c(-2, -1, 0, 0.5, 1))
}
\arguments{
\item{Z}{M by K matrix of summary Z-statistics for M variants across K traits}

\item{Sig}{estimated trait correlation matrix (K by K)}

\item{R}{variant LD correlation matrix (M by M)}

\item{rho}{weight assigned to the ET}
}
\value{
\describe{
  \item{p.value}{ test p-values for: AT, VC, ET }
  \item{pval}{ vector of all p-values }
  \item{rho.est}{ the optimal rho weight }
}
}
\description{
We use the LD score regression (Bulik-Sullivan et al.) to accurately estimate the trait correlation,
which is then used to construct the multi-trait association tests of multiple variants in a gene or pathway: PC based test (ET), 
variance components test (VT), and their adaptive test (AT)
}
\examples{
K = 4; M = 20
R = cor(matrix(rnorm(500*M),500,M)*sqrt(0.8)+rnorm(500)*sqrt(0.2))
Y = matrix(rnorm(100*K), 100,K)*sqrt(0.8)+rnorm(100)*sqrt(0.2)
Y[,1] = -Y[,1]; Sig = cor(Y)
Z = matrix(rnorm(K*M),M,K)
js = sample(1:M, size=round(M*0.4))
Z[js,] = Z[js,] + 1.5
EMSATS(Z,Sig,R)
Z = matrix(rnorm(K*M),M,K)
ij = cbind(sample(1:M, size=20, rep=TRUE), sample(1:K, size=20, rep=TRUE))
Z[ij] = Z[ij] + rnorm(20)*1.5
EMSATS(Z,Sig,R)
}
\references{
Guo,B. and Wu,B. (2017) PCA based adaptive multi-trait SNP-set association test using the GWAS summary statistics. tech rep.
}
