% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgamut.R
\name{fgamut}
\alias{fgamut}
\title{Fast implementation of GAMuT (Gene Association with Multiple Traits)}
\usage{
fgamut(Y, G, X = NULL, W = NULL, W.beta = c(1, 25))
}
\arguments{
\item{Y}{outcome matrix. sample in rows.}

\item{G}{genotype matrix. sample in rows.}

\item{X}{covariate matrix to be adjusted. sample in rows.}

\item{W}{variant weight; use Beta dist weight if NULL}

\item{W.beta}{Beta distribution parameters for the variant weight. Default to Beta(1,25).}
}
\value{
a p-value vector p.value with four components
\describe{
  \item{proj}{ p-value for the GAMuT with projection matrix kernel }
  \item{Cproj}{ corrected p-value for the GAMuT with projection matrix kernel }
  \item{linear}{ p-value for the GAMuT with linear kernel }
  \item{Clinear}{ corrected p-value for the GAMuT with linear kernel }
}
}
\description{
Return the multi-trait variant set association test p-values from the Broadaway et. al (2016) approach.
 Both tests from linear kernel and projection matrix are computed.
This is a much faster implementation following the approach of Wu and Pankow (2016) using the multiple
linear regression based score test. We also provide modified (and more accurate) p-values following the approach of
Wu et al. (2016).
}
\references{
Broadaway, K.A., Cutler, D.J., Duncan, R., Moore, J.L., Ware, E.B., Jhun, M.A., Bielak, L.F., Zhao, W., Smith, J.A., Peyser, P.A., Kardia, S.L.R., Ghosh, D., Epstein, M.P. (2016). A Statistical Approach for Testing Cross-Phenotype Effects of Rare Variants. \emph{The American Journal of Human Genetics}, 98, 525--540.

Wu,B. and Pankow,J.S. (2016) Sequence kernel association test of multiple continuous phenotypes. \emph{Genetic Epidemiology}, 40(2), 91-100.

Wu,B., Guan,W., Pankow,J.S. (2016) On efficient and accurate calculation of significance p-values for sequence kernel association test of variant set. \emph{Annals of human genetics}, 80(2), 123-135.

Wu,B. and Qin,H. (2018). On testing cross-phenotype effects of rare variants. \emph{AHG}, under revision. \url{http://www.umn.edu/~baolin/research/gamut}
}
