% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamut.R
\name{gamut}
\alias{gamut}
\title{GAMuT: Gene Association with Multiple Traits}
\usage{
gamut(PHENO, GENO, COVAR = NULL, W = NULL, W.beta = c(1, 25))
}
\arguments{
\item{PHENO}{outcome matrix. sample in rows.}

\item{GENO}{genotype matrix. sample in rows.}

\item{COVAR}{covariate matrix to be adjusted. sample in rows.}

\item{W}{variant weight; use Beta dist weight if NULL}

\item{W.beta}{Beta distribution parameters for the variant weight. Default to Beta(1,25).}
}
\value{
a p-value vector p.value with two components
\describe{
  \item{proj}{ p-value for GAMuT with the projection matrix kernel }
  \item{linear}{ p-value for GAMuT with the linear kernel }
}
}
\description{
Return the multi-trait variant set association test p-values from the Broadaway et. al (2016) approach.
 Both tests from linear kernel and projection matrix are computed.
}
\references{
Broadaway, K.A., Cutler, D.J., Duncan, R., Moore, J.L., Ware, E.B., Jhun, M.A., Bielak, L.F., Zhao, W., Smith, J.A., Peyser, P.A., Kardia, S.L.R., Ghosh, D., Epstein, M.P. (2016). A Statistical Approach for Testing Cross-Phenotype Effects of Rare Variants. \emph{The American Journal of Human Genetics}, 98, 525--540.

Original R codes posted at \url{http://genetics.emory.edu/labs/epstein/software/gamut}
}
