% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mskat.R
\name{KAT.pval}
\alias{KAT.pval}
\title{Compute the tail probability of 1-DF chi-square mixtures}
\usage{
KAT.pval(Q.all, lambda, acc = 1e-09, lim = 1e+06)
}
\arguments{
\item{Q.all}{test statistics}

\item{lambda}{mixing coefficients}

\item{acc}{error bound}

\item{lim}{maximum number of integration terms}
}
\description{
Use Davies' method; if fail, switch to the Saddlepoint approx
}
