% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msatZ-list.R
\name{Lmsatz}
\alias{Lmsatz}
\title{Compute the list of significant SNPs using the GWAS summary data based multi-trait association tests}
\usage{
Lmsatz(Z, Sig, alpha = 5e-08)
}
\arguments{
\item{Z}{matrix of summary Z-statistics (SNPs by traits)}

\item{Sig}{the estimated marginal trait correlation matrix}

\item{alpha}{desired genome-wide significance level (default to 5E-8)}
}
\value{
\describe{
  \item{idM}{ significant SNP list for minP }
  \item{idS}{ significant SNP list for SZ }
  \item{idS2}{ significant SNP list for SZ2 }
}
}
\description{
We compute the significant SNPs for three tests: minimum marginal test p-value (minP); SZ test; SZ2 test.
}
\references{
Guo,B. and Wu,B. (2017) GWAS summary data based multi-trait SNP-set association test to identify novel genetic variants. tech rep.
}
