% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvkm.R
\name{MKMR}
\alias{MKMR}
\title{Multivariate Kernel Machine Regression test of rare variant set}
\usage{
MKMR(Y, X.list, Sigma = NULL, Gm, W = NULL, W.beta = c(1, 25),
  n.sim = NULL)
}
\arguments{
\item{Y}{pxn matrix, each row is one response, each column is one subject}

\item{X.list}{a list where each item of list is a qxn matricx (parametric design matrix),
where each row is one covariate (possiblly different for all items in Y), intercept must be included!}

\item{Sigma}{covariance matrix of multivariate outcomes; estimate from data if NULL}

\item{Gm}{genotype matrix with SNPs in column}

\item{W}{variant weight; use Beta dist weight if NULL}

\item{W.beta}{Beta distribution parameters for the variant weight. Default to Beta(1,25).}

\item{n.sim}{number of simulation to obtain the null distribution of the test statistic; using analytical method if NULL}
}
\description{
Return the variant set association test statistic and p-value.
  Aadapted from Maity et. al approach for rare variant set association test using weighted linear kernel.
}
\references{
Maity,A., Sullivan,P.F., Tzeng,J. (2012) Multivariate Phenotype Association Analysis by Marker-Set Kernel Machine Regression. Genet. Epidemiol. 36, 686–695.
}
