% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msatZ.R
\name{msatz}
\alias{msatz}
\title{Multi-trait SNP-set association test using GWAS summary data}
\usage{
msatz(Z, Sig, R, rho = 0:10/10)
}
\arguments{
\item{Z}{M by K matrix of summary Z-statistics for M variants across K traits}

\item{Sig}{estimated trait correlation matrix (K by K)}

\item{R}{variant LD correlation matrix (M by M)}

\item{rho}{weight assigned to the ST}
}
\value{
\describe{
  \item{p.value}{ test p-values for: AT, SST, ST }
  \item{pval}{ vector of all p-values }
  \item{rho.est}{ the optimal rho weight }
}
}
\description{
Efficient multi-trait SNP-set association tests: squared sum test (SST), sum test (ST), their adaptive test (AT)
}
\examples{
K = 4; M = 20
R = cor(matrix(rnorm(500*M),500,M)*sqrt(0.8)+rnorm(500)*sqrt(0.2))
Y = matrix(rnorm(100*K), 100,K)*sqrt(0.8)+rnorm(100)*sqrt(0.2)
Y[,1] = -Y[,1]; Sig = cor(Y)
Z = matrix(rnorm(K*M),M,K)
js = sample(1:M, size=round(M*0.4))
Z[js,] = Z[js,] + 1.5
msatz(Z,Sig,R)
Z = matrix(rnorm(K*M),M,K)
ij = cbind(sample(1:M, size=20, rep=TRUE), sample(1:K, size=20, rep=TRUE))
Z[ij] = Z[ij] + rnorm(20)*1.5
msatz(Z,Sig,R)
}
\references{
Guo,B. and Wu,B. (2017) GWAS summary data based multi-trait SNP-set association test to identify novel genetic variants. tech rep.
}
