% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mskat.R
\name{MSKAT.cnull}
\alias{MSKAT.cnull}
\title{Fit a null multi-trait regression model}
\usage{
MSKAT.cnull(Ys, X = NULL)
}
\arguments{
\item{Ys}{matrix of continuous outcomes with variables in column}

\item{X}{matrix of covariates to be adjusted}
}
\description{
Compute a null multivariate regression model
}
