% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mskat.R
\name{MSKAT}
\alias{MSKAT}
\title{Multi-trait sequence kernel association test of variant set}
\usage{
MSKAT(obj, G, W = NULL, W.beta = c(1, 25))
}
\arguments{
\item{obj}{fitted null model from MSKAT.cnull}

\item{G}{genotype matrix with SNPs in column}

\item{W}{variant weight; use Beta dist weight if NULL}

\item{W.beta}{Beta distribution parameters for the variant weight}
}
\value{
\describe{
  \item{p.value}{ p-values for the two kernel statistics (Q and Q'; see ref)}
  \item{Q}{ test statistics (Q and Q'; see ref)}
}
}
\description{
Compute the variant set association test statistic and p-value
}
\examples{
Y = matrix(rnorm(2000), 1000,2)
X = matrix(rnorm(2000), 1000,2)
G = matrix(rbinom(10000,2,0.02), 1000,10)
MSKAT(MSKAT.cnull(Y,X), G, W.beta=c(1,25))
## time-consuming; not run
## X.list = vector('list',2); for(i in 1:2) X.list[[i]] = rbind(1, t(X))
## MKMR(t(Y),X.list, Gm=G, W.beta=c(1,25))
}
\references{
Wu,B. and Pankow,J.S. (2016) Sequence kernel association test of multiple continuous phenotypes. Genetic Epidemiology, 40(2), 91-100.
}
