% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqf_davies.R
\name{Wdavies}
\alias{Wdavies}
\title{Compute the tail prob for weighted sum of 1-DF chi-square random variables}
\usage{
Wdavies(qx, lambda, delta = NULL, lim = 1e+09, acc = 1e-06)
}
\arguments{
\item{qx}{quantile to compute tail prob}

\item{lambda}{weights for 1-DF chi-square rvs}

\item{delta}{non-centrality parameters for 1-DF chi-square rvs. Default to zeros.}

\item{lim}{total number of integration terms}

\item{acc}{accuracy for computed probability}
}
\value{
\describe{
 \item{Qq}{ tail prob }
 \item{trace}{vector, indicating performance of procedure, with the following components: 1: absolute value sum, 2: total number of integration terms, 3: number of integrations, 4: integration interval in main integration, 5: truncation point in initial integration, 6: standard deviation of convergence factor term, 7: number of cycles to locate integration parameters}
 \item{ifault}{ fault indicator: 0: no error, 1: requested accuracy could not be obtained, 2: round-off error possibly significant, 3: invalid parameters, 4: unable to locate integration parameters }
}
}
\description{
Imported from CompQuadForm R package (version 1.4.3; 10/4/2017). Slight simplifications for sum of central chi-squares.
}
